First review by vk (#43)
* draft

* draft of numericall section

* draft to jump back to numeric

* A good writeup up to purely random analytically

* write up of purely random strategies

This includes the analytical and the numerical section

* reactive analytically

* fix reactive script

* finish algorithm reactive

* reactive numerics sections

* stability of defection

* first draft of paper + reactive script

* morning corrections

* Reactive script (#31)

* rename functions and add docstrings

* tweak implementation and add tests

* Fix script for matches

Find optimal reactive for matches and train gambler

* implement tournaments

* fix path to main.py

* fix stability of defection section

* generate images for reactive players

* Fix reactive script (#32)

* remove function from script.

All functions used in our main script leaved in the same file, main.py.
The functions have been moved to optimisation so they can be used
in different scripts or just call them via opt_mo

* no longer need to import main just opt_mo

* fix reactive script

Now utility is being optimised using bayesian algorithm as well.
grambler is notbeing trained cause it is not needed.

* change title

* minor changes

* fix comments in introduction.

Apart from adding Nowak's article (opened an issue).

* change title of section two

* fix utility section

* fix reactive implementation. (#39)

After performing several numerical trials we noticed that the reactive
optimisation was not captured best behaviour. This fixes the issue.

The issue was that when p1 is 0 or 1 then we solve only the second
partial derivative. The derivative which is depended on p2

* dummy

* Last tweaks in tility against memory one players

Fixes plots, Theorem 2 and discuss the validation
though Stewart and Plotkin's tournament. Table
with results still need to be added.

* add reslts against stewart and plotkin's tour

* re write best response section

This section has been decicaded to best responses
against a group and a signle memory one strategy.

The structure (for most) has been discussion, lemma, proof
and empirical results.

Things to consider are:

- concavity. Is the sum of non concave functions always non concave
- proofs. Some lemmas follow from a lemma with a proof. Do I need
to add proofs for these as well.

* draft of stability of defection

* tweak previous section for stability section

* fix stability of defection section.

* draft of optimisation of memory one strategies

* Second draft.

This needs morning coffee and corrections.

* corrections to second draft

* Edit introductions

- Minor white space adjustements
- A few rewordings
- Adding a todo to include a reference.

* Edit section on formulation

- Various small edits and TODOs
- One major TODO: the figure that was meant to go with thetournament
  theorem is not at all the correct one.

* Edit section on best responses

- Added a number of TODOs
- Some minor edits
- Reword the best response theorem.
- Add a suggestion for a table of some sort

* Edit the results section.

I will re read this properly once we've spoken together.

* Edit stability and optimisation section.

- Minor changes
- Some suggestions

* Edit numerical results section.

* Add comment about moving results to appendix.